src/requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
fastapi==0.111.0
gpt4all==2.7.0
langchain-chroma==0.1.1
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.62
langchain-huggingface==0.0.3
langchain-mistralai==0.1.8
langchain-openai==0.1.10
langchainhub==0.1.20
langgraph==0.1.1
langserve==0.2.2
llama_cpp_python==0.2.79
python-dotenv==1.0.1
pypdf==4.2.0
pytube==15.0.0
sse-starlette==2.1.2
youtube-transcript-api==0.6.2
